[HUBS] Refleciton probes (#59)
We would like for scenes to be able to define multiple environment maps to make baked lighting more dynamic. This patch adds a new primitive called **ReflectionProbe**. A **ReflectionProbe** has a `box` which defines the axis aligned bounding box in which it applies, and a `texture` which defines the equirectangular texture that will be used as the environment map for objects inside this box. If an object is overlapping multiple **ReflectionProbe** boxes its environment map will be a mix of the 2 most overlapped boxes, with the mixing being proportional to the amount of overlap. Object overlap is computed by the object's axis aligned bounding box.

By default objects will not be effected by reflection probes. An object must set the property `reflectionProbeType` to either `static` or `dynamic`. When `static` the object will only check itself against **ReflectionProbes** the first time it is rendered. When `dynamic` this check will happen every frame. The checks should be fairly lightweight, but for static geometry like scenes it is a nice optimization. Materials that explicitly define `envMap` will override any reflection probes that would have applied to this material.

This implementation should be considered experimental and will likely go through several more rounds of changes