[HUBS] Raise childrenNeedMatrixWorldUpdate only if needed
childrenNeedMatrixWorldUpdate = true can have a big bad
performance impact because it can cause all the child world
matrices update.

So it would be worth to raise the flag in updateMatrices()
only if the new world matrix values have diff from the current
values.

Checking the diff can be performance panalty for moving objects
but generally not all the objects in a scene move at a certain
frame. Probably the benefit wins to the penalty in most cases.